requirements-dev.txt
# Development dependencies
-r requirements.txt

# Development tools
pytest==6.2.5
pytest-django==4.5.2
black==21.12b0
flake8==4.0.1
factory-boy==2.12.0
